+++ Kite-SDK/PSPrintSDK/OLAnalytics.m
@@ -34,12 +34,9 @@
 #include <sys/sysctl.h>
 #import "OLKiteABTesting.h"
 #import "NSDictionary+RequestParameterData.h"
-#import "OLUserSession.h"
 #import "OLKiteUtils.h"
 
 static NSString *const kKeyUserDistinctId = @"ly.kite.sdk.kKeyUserDistinctId";
-static NSString *const kOLMixpanelToken = @"cdf64507670dd359c43aa8895fb87676";
-static NSString *const kOLMixpanelURL = @"https://api.mixpanel.com/track/";
 static NSString *const kKeySDKLaunchCount = @"ly.kite.sdk.kKeySDKLaunchCount";
 static NSString *const kKeyOLDevicePushToken = @"ly.kite.sdk.kKeyOLDevicePushToken";
 
@@ -53,11 +50,6 @@ static NSString *nonNilStr(NSString *str) {
 
 static BOOL optInToRemoteAnalytics = NO;
 
-@interface OLProduct (Private)
-- (NSDecimalNumber*) unitCostDecimalNumber;
-- (NSString *)currencyCode;
-@end
-
 @interface OLKitePrintSDK (Private)
 
 + (NSString *)apiEndpoint;
@@ -65,7 +57,20 @@ static BOOL optInToRemoteAnalytics = NO;
 
 @end
 
+@interface OLAnalytics () <AnalyticsDelegate>
+@end
+
 @implementation OLAnalytics
+
++ (instancetype)sharedInstance {
+    static dispatch_once_t once;
+    static OLAnalytics * sharedInstance;
+    dispatch_once(&once, ^ {
+        sharedInstance = [[self alloc] init];
+        [Analytics shared].delegate = sharedInstance;
+    });
+    return sharedInstance;
+}
 
 + (void)incrementLaunchSDKCount{
     NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
@@ -82,6 +87,7 @@ static BOOL optInToRemoteAnalytics = NO;
 
 + (void)setOptInToRemoteAnalytics:(BOOL)optIn {
     optInToRemoteAnalytics = optIn;
+    [Analytics shared].optInToRemoteAnalytics = optIn;
 }
 
 + (NSString *)userDistinctId{
@@ -179,86 +185,6 @@ static BOOL optInToRemoteAnalytics = NO;
     [session finishTasksAndInvalidate];
 }
 
-+ (void)sendToMixPanelWithDictionary:(NSDictionary *)dict{
-    if (!optInToRemoteAnalytics) {
-        return;
-    }
-    
-    NSError *error;
-    NSData * jsonData = [NSJSONSerialization dataWithJSONObject:dict options:NSJSONWritingPrettyPrinted error:&error];
-    
-    NSDictionary *params = @{@"ip": @"1",@"data" : [jsonData base64EncodedStringWithOptions:0]};
-    
-    NSURL *baseURL = [NSURL URLWithString:kOLMixpanelURL];
-    
-    NSMutableURLRequest *request = [[NSMutableURLRequest alloc] initWithURL:baseURL];
-    request.HTTPMethod = @"POST";
-    request.HTTPBody = [params requestParameterData];
-    [request addValue:@"application/json" forHTTPHeaderField:@"Content-Type"];
-    [request addValue:@"application/json" forHTTPHeaderField:@"Accept"];
-    
-    NSURLSessionConfiguration *sessionConfig = [NSURLSessionConfiguration defaultSessionConfiguration];
-    
-    NSURLSession *session = [NSURLSession sessionWithConfiguration:sessionConfig
-                                                          delegate:nil
-                                                     delegateQueue:nil];
-    [[session dataTaskWithRequest:request] resume];
-    [session finishTasksAndInvalidate];
-}
-
-+ (NSDictionary *)defaultDictionaryForEventName:(NSString *)eventName{
-    NSString *environment = @"Live";
-    if ([OLKitePrintSDK environment] == OLKitePrintSDKEnvironmentSandbox) {
-        environment = @"Sandbox";
-    }
-    NSDictionary *info = [[NSBundle mainBundle] infoDictionary];
-    NSString *bundleName = nil;
-    if ([info objectForKey:@"CFBundleDisplayName"] == nil) {
-        bundleName = [[[NSBundle mainBundle] infoDictionary] objectForKey:(NSString *) kCFBundleNameKey];
-    } else {
-        bundleName = [NSString stringWithFormat:@"%@", [info objectForKey:@"CFBundleDisplayName"]];
-    }
-    NSString *appVersion = [[NSBundle mainBundle] objectForInfoDictionaryKey:@"CFBundleVersion"];
-    NSString *apiKey = [OLKitePrintSDK apiKey] == nil ? @"Unknown" : [OLKitePrintSDK apiKey];
-    
-    NSString *localeCountry = [NSLocale currentLocale].localeIdentifier;
-    
-    NSMutableDictionary *propertiesDict = [@{
-                                             @"token" : kOLMixpanelToken,
-                                             @"distinct_id" : [OLAnalytics userDistinctId],
-                                             @"App Bundle Id" : [[NSBundle mainBundle] bundleIdentifier],
-                                             @"App Name" : bundleName,
-                                             @"App Version" : appVersion,
-                                             @"platform" : @"iOS",
-                                             @"platform version" : [[UIDevice currentDevice] systemVersion],
-                                             @"model" : [OLAnalytics platform],
-                                             @"Screen Height" : @([UIScreen mainScreen].bounds.size.height),
-                                             @"Screen Width" : @([UIScreen mainScreen].bounds.size.width),
-                                             @"Environment" : environment,
-                                             @"API Key": apiKey,
-                                             @"Kite SDK Version": kOLKiteSDKVersion,
-                                             @"Locale Country": localeCountry,
-                                             @"Progressive Template Loading" : @([OLKiteABTesting sharedInstance].progressiveTemplateLoading)
-                                             } mutableCopy];
-    if ([extraInfo isKindOfClass:[NSDictionary class]]){
-        for (id key in extraInfo.allKeys){
-            if ([key isKindOfClass:[NSString class]]){
-                propertiesDict[[@"App Extra-" stringByAppendingString:key]] = extraInfo[key];
-            }
-        }
-    }
-    
-    NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
-    NSNumber *n = [defaults objectForKey:kKeySDKLaunchCount];
-    if (n){
-        propertiesDict[@"SDK Launch Count"] = n;
-    }
-    
-    NSDictionary *dict = @{@"event" : eventName,
-                           @"properties" : propertiesDict};
-    return dict;
-}
-
 + (NSDictionary *)dictionaryForJob:(id<OLPrintJob>)job{
     NSMutableDictionary *dict = [[NSMutableDictionary alloc] init];
     OLProductTemplate *template = [OLProductTemplate templateWithId:[job templateId]];
@@ -271,334 +197,68 @@ static BOOL optInToRemoteAnalytics = NO;
     return dict;
 }
 
-//+ (void)reportAnalyticsEventToDelegate:(NSString *)eventName
-//                                   job:(id<OLPrintJob>)job
-//                            printOrder:(OLPrintOrder *)order
-//                             extraInfo:(NSDictionary *)info{
-//    if ([[OLUserSession currentSession].kiteVc.delegate respondsToSelector:@selector(logKiteAnalyticsEventWithInfo:)]){
-//        NSMutableDictionary *dict = [[NSMutableDictionary alloc] init];
-//        dict[kOLAnalyticsEventName] = nonNilStr(eventName);
-//        if (job){
-//            [dict addEntriesFromDictionary:[OLAnalytics dictionaryForJob:job]];
-//        }
-//        if (info){
-//            [dict addEntriesFromDictionary:info];
-//        }
-//        if (order){
-//            NSMutableArray *items = [[NSMutableArray alloc] init];
-//            for (id<OLPrintJob> job in order.jobs){
-//                [items addObject:[OLAnalytics dictionaryForJob:job]];
-//            }
-//            dict[kOLAnalyticsItemsInOrder] = items;
-//
-//            dict[kOLAnalyticsNumberOfPhotosInOrder] = [NSNumber numberWithInteger:order.totalAssetsToUpload];
-//            dict[kOLAnalyticsPromoCode] = order.promoCode;
-//            dict[kOLAnalyticsCurrencyCode] = order.currencyCode;
-//            dict[kOLAnalyticsOrderId] = order.receipt;
-//            [order costWithCompletionHandler:^(OLPrintOrderCost *cost, NSError *error){
-//                dict[kOLAnalyticsOrderCost] = [cost totalCostInCurrency:order.currencyCode];
-//                dict[kOLAnalyticsOrderShippingCost] = [cost shippingCostInCurrency:order.currencyCode];
-//
-//                [[OLUserSession currentSession].kiteVc.delegate logKiteAnalyticsEventWithInfo:dict];
-//            }];
-//        }
-//        else{
-//            [[OLUserSession currentSession].kiteVc.delegate logKiteAnalyticsEventWithInfo:dict];
-//        }
-//    }
-//}
++ (void)reportAnalyticsEventToDelegate:(NSString *)eventName job:(id<OLPrintJob>)job extraInfo:(NSDictionary *)info{
+    if (![OLAnalytics sharedInstance].delegate){
+        return;
+    }
+    
+    NSMutableDictionary *dict = [[NSMutableDictionary alloc] init];
+    dict[kOLAnalyticsEventName] = nonNilStr(eventName);
+    if (job){
+        [dict addEntriesFromDictionary:[OLAnalytics dictionaryForJob:job]];
+    }
+    if (info){
+        [dict addEntriesFromDictionary:info];
+    }
+    
+    [[OLAnalytics sharedInstance].delegate logKiteAnalyticsEventWithInfo:dict];
+}
 
-#pragma mark Track Main Events - Sent to MixPanel
+#pragma mark Track Events
 
 + (void)trackKiteViewControllerLoadedWithEntryPoint:(NSString *)entryPoint {
     [OLAnalytics incrementLaunchSDKCount];
     
     NSString *eventName = kOLAnalyticsEventNameKiteLoaded;
-    NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
-    [dict[@"properties"] setObject:nonNilStr(entryPoint) forKey:@"Entry Point"];
-    [OLAnalytics sendToMixPanelWithDictionary:dict];
-    
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsEventLevel : @1}];
+    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil extraInfo:@{kOLAnalyticsEventType : kOLAnalyticsEventTypeAction,}];
 }
 
 + (void)trackCategoryListScreenViewed{
     NSString *eventName = kOLAnalyticsEventNameCategoryListScreenViewed;
-    NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
-    [dict[@"properties"] setObject:nonNilStr([OLKiteABTesting sharedInstance].qualityBannerType) forKey:@"Quality Banner Type"];
-    [dict[@"properties"] setObject:nonNilStr([OLKiteABTesting sharedInstance].productTileStyle) forKey:@"Product Tile Style"];
-    [dict[@"properties"] setObject:nonNilStr([OLKiteABTesting sharedInstance].promoBannerText) forKey:@"Promo Banner Variant"];
-    [OLAnalytics sendToMixPanelWithDictionary:dict];
-    
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsEventLevel : @1}];
+    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil extraInfo:@{kOLAnalyticsEventType : kOLAnalyticsEventTypeScreenViewed,}];
 }
 
 + (void)trackQualityInfoScreenViewed{
     NSString *eventName = kOLAnalyticsEventNameQualityInfoScreenViewed;
-    NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
-    [dict[@"properties"] setObject:[OLKiteABTesting sharedInstance].qualityBannerType forKey:@"Quality Banner Type"];
-    [OLAnalytics sendToMixPanelWithDictionary:dict];
     
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsEventLevel : @1}];
+    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil extraInfo:@{kOLAnalyticsEventType : kOLAnalyticsEventTypeScreenViewed,}];
 }
 
 + (void)trackPrintAtHomeTapped{
     NSString *eventName = kOLAnalyticsEventNamePrintAtHomeTapped;
-    NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
-    [OLAnalytics sendToMixPanelWithDictionary:dict];
     
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsEventLevel : @1}];
+    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil extraInfo:@{}];
 }
 
 + (void)trackProductDetailsScreenViewed:(OLProductTemplate *)productTemplate hidePrice:(BOOL)hidePrice{
     NSString *eventName = kOLAnalyticsEventNameProductDetailsScreenViewed;
-    NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
-    [dict[@"properties"] setObject:nonNilStr(productTemplate.name) forKey:kOLAnalyticsProductName];
-    [dict[@"properties"] setObject:nonNilStr(productTemplate.identifier) forKey:kOLAnalyticsProductId];
-    [dict[@"properties"] setObject:hidePrice ? @"YES" : @"NO" forKey:@"Hide Price on Product Description"];
-    [OLAnalytics sendToMixPanelWithDictionary:dict];
-    
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : nonNilStr(productTemplate.name), kOLAnalyticsProductId : nonNilStr(productTemplate.identifier), kOLAnalyticsEventLevel : @1}];
+    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil extraInfo:@{kOLAnalyticsProductName : nonNilStr(productTemplate.name), kOLAnalyticsProductId : nonNilStr(productTemplate.identifier), kOLAnalyticsEventType : kOLAnalyticsEventTypeScreenViewed,}];
 }
 
 + (void)trackProductListScreenViewedWithTemplateClass:(NSString *)templateClassString{
     NSString *eventName = kOLAnalyticsEventNameProductListScreenViewed;
-    NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
-    [dict[@"properties"] setObject:nonNilStr(templateClassString) forKey:@"Product Class"];
-    [OLAnalytics sendToMixPanelWithDictionary:dict];
-    
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductCategory : templateClassString, kOLAnalyticsEventLevel : @1}];
+    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil extraInfo:@{kOLAnalyticsProductCategory : templateClassString, kOLAnalyticsEventType : kOLAnalyticsEventTypeScreenViewed,}];
 }
 
 + (void)trackImagePickerScreenViewed:(NSString *)productName{
     NSString *eventName = kOLAnalyticsEventNameImagePickerScreenViewed;
-    NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
-    [dict[@"properties"] setObject:productName forKey:@"Product Name"];
-    [OLAnalytics sendToMixPanelWithDictionary:dict];
-    
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventLevel : @1}];
+    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventType : kOLAnalyticsEventTypeScreenViewed,}];
 }
 
 + (void)trackReviewScreenViewed:(NSString *)productName{
     NSString *eventName = kOLAnalyticsEventNameReviewScreenViewed;
-    NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
-    [dict[@"properties"] setObject:productName forKey:@"Product Name"];
-    [OLAnalytics sendToMixPanelWithDictionary:dict];
-    
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventLevel : @1}];
+    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventType : kOLAnalyticsEventTypeScreenViewed,}];
 }
-
-+ (void)trackKiteDismissed{
-    NSString *eventName = kOLAnalyticsEventNameKiteDismissed;
-    
-    NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
-    [OLAnalytics sendToMixPanelWithDictionary:dict];
-    
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsEventLevel : @1}];
-}
-
-+ (void)trackBasketIconTappedWithNumberBadged:(NSInteger)number{
-    NSString *eventName = kOLAnalyticsEventNameBasketIconTapped;
-    
-    NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
-    [dict[@"properties"] setObject:[NSNumber numberWithInteger:number] forKey:kOLAnalyticsNumberOnBadge];
-    [OLAnalytics sendToMixPanelWithDictionary:dict];
-    
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsNumberOnBadge : [NSNumber numberWithInteger:number], kOLAnalyticsEventLevel : @1}];
-}
-
-#pragma mark Track Secondary Events - Not Sent to MixPanel
-
-+ (void)trackProductListScreenHitBackTemplateClass:(NSString *)templateClassString{
-//    [OLAnalytics reportAnalyticsEventToDelegate:kOLAnalyticsEventNameProductListScreenHitBack job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductCategory : templateClassString, kOLAnalyticsEventLevel : @2}];
-}
-
-+ (void)trackPhotoProviderPicked:(NSString *)provider forProductName:(NSString *)productName{
-    NSString *eventName = kOLAnalyticsEventNameKitePhotoProviderSelected;
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : nonNilStr(productName), kOLAnalyticsPhotoSource : nonNilStr(provider), kOLAnalyticsEventLevel : @2}];
-}
-
-+ (void)trackQualityScreenHitBack{
-//    [OLAnalytics reportAnalyticsEventToDelegate:kOLAnalyticsEventNameKiteQualityInfoScreenHitBack job:nil printOrder:nil extraInfo:@{kOLAnalyticsEventLevel : @2}];
-}
-
-+ (void)trackOrderHistoryScreenViewed{
-//    [OLAnalytics reportAnalyticsEventToDelegate:kOLAnalyticsEventNameOrderHistoryViewed job:nil printOrder:nil extraInfo:@{kOLAnalyticsEventLevel : @2}];
-}
-
-+ (void)trackOrderHistoryScreenDismissed{
-//    [OLAnalytics reportAnalyticsEventToDelegate:kOLAnalyticsEventNameOrderHistoryDismissed job:nil printOrder:nil extraInfo:@{kOLAnalyticsEventLevel : @2}];
-}
-
-//+ (void)trackBasketItemQtyDownForItem:(id<OLPrintJob>)item inOrder:(OLPrintOrder *)printOrder applePayIsAvailable:(NSString *)applePayIsAvailable{
-//    NSString *eventName = kOLAnalyticsEventNameBasketItemQuantityDown;
-////    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:item printOrder:printOrder extraInfo:@{kOLAnalyticsEventLevel : @2}];
-//}
-
-+ (void)trackReviewScreenHitBack:(NSString *)productName numberOfPhotos:(NSInteger)numberOfPhotos{
-    NSString *eventName = kOLAnalyticsEventNameReviewScreenHitBack;
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsNumberOfPhotos : [NSNumber numberWithInteger:numberOfPhotos], kOLAnalyticsEventLevel : @2}];
-}
-
-//+ (void)trackBasketItemQtyUpForItem:(id<OLPrintJob>)item inOrder:(OLPrintOrder *)printOrder applePayIsAvailable:(NSString *)applePayIsAvailable{
-//    NSString *eventName = kOLAnalyticsEventNameBasketItemQuantityUp;
-////    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:item printOrder:printOrder extraInfo:@{kOLAnalyticsEventLevel : @2}];
-//}
-
-+ (void)trackEditScreenFinishedEditingPhotoForProductName:(NSString *)productName{
-    NSString *eventName = kOLAnalyticsEventNameEditScreenFinishedEditingPhoto;
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventLevel : @2}];
-}
-
-+ (void)trackEditPhotoTappedForProductName:(NSString *)productName{
-    NSString *eventName = kOLAnalyticsEventNameEditPhotoTapped;
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventLevel : @2}];
-}
-
-+ (void)trackReviewScreenDecrementedPhotoQtyForProductName:(NSString *)productName{
-    NSString *eventName = kOLAnalyticsEventNameReviewScreenDecrementedPhotoQuantity;
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventLevel : @2}];
-}
-
-//+ (void)trackBasketScreenDidDeleteItem:(id<OLPrintJob>)item inOrder:(OLPrintOrder *)printOrder applePayIsAvailable:(NSString *)applePayIsAvailable{
-//    NSString *eventName = kOLAnalyticsEventNameBasketScreenDeleteItem;
-////    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:item printOrder:printOrder extraInfo:@{kOLAnalyticsEventLevel : @2}];
-//}
-
-//+ (void)trackBasketScreenHitEditItemDone:(id<OLPrintJob>)item inOrder:(OLPrintOrder *)printOrder applePayIsAvailable:(NSString *)applePayIsAvailable{
-//    NSString *eventName = kOLAnalyticsEventNameBasketScreenDidEditItem;
-////    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:item printOrder:printOrder extraInfo:@{kOLAnalyticsEventLevel : @2}];
-//}
-
-+ (void)trackProductDetailsViewOpened:(NSString *)productName hidePrice:(BOOL)hidePrice{
-    NSString *eventName = kOLAnalyticsEventNameProductDetailsOpened;
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventLevel : @2}];
-}
-
-+ (void)trackProductDetailsViewClosed:(NSString *)productName hidePrice:(BOOL)hidePrice{
-    NSString *eventName = kOLAnalyticsEventNameProductDetailsClosed;
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventLevel : @2}];
-}
-
-//+ (void)trackBasketScreenHitBackForOrder:(OLPrintOrder *)printOrder applePayIsAvailable:(NSString *)applePayIsAvailable{
-//    NSString *eventName = kOLAnalyticsEventNameBasketScreenHitBack;
-////    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:printOrder extraInfo:@{kOLAnalyticsEventLevel : @2}];
-//}
-//
-//+ (void)trackBasketScreenSuccessfullyAppliedPromoCode:(NSString *)code forOrder:(OLPrintOrder *)order{
-////    [OLAnalytics reportAnalyticsEventToDelegate:kOLAnalyticsEventNameBasketScreenSuccessfullyAppliedPromoCode job:nil printOrder:order extraInfo:@{kOLAnalyticsPromoCode : nonNilStr(code), kOLAnalyticsEventLevel : @2}];
-//}
-//
-//+ (void)trackBasketScreenUnsuccessfullyAppliedPromoCode:(NSString *)code withError:(NSError *)error forOrder:(OLPrintOrder *)order{
-////    [OLAnalytics reportAnalyticsEventToDelegate:kOLAnalyticsEventNameBasketScreenPromoCodeFailed job:nil printOrder:order extraInfo:@{kOLAnalyticsPromoCode : nonNilStr(code), kOLAnalyticsError : error ? error : @"", kOLAnalyticsEventLevel : @2}];
-//}
-//
-//+ (void)trackBasketScreenDidTapOnPromoCodeBoxforOrder:(OLPrintOrder *)order{
-////    [OLAnalytics reportAnalyticsEventToDelegate:kOLAnalyticsEventNameBasketScreenDidTaponPromoCodeBox job:nil printOrder:order extraInfo:@{kOLAnalyticsEventLevel : @2}];
-//}
-//
-//+ (void)trackBasketScreenPaymentMethodDidCancel:(NSString *)method forOrder:(OLPrintOrder *)order applePayIsAvailable:(NSString *)applePayIsAvailable{
-////    [OLAnalytics reportAnalyticsEventToDelegate:kOLAnalyticsEventNameBasketScreenPaymentMethodHitBack job:nil printOrder:order extraInfo:@{kOLAnalyticsPaymentMethod : method, kOLAnalyticsEventLevel : @2}];
-//}
-
-+ (void)trackReviewScreenIncrementedPhotoQtyForProductName:(NSString *)productName{
-    NSString *eventName = kOLAnalyticsEventNameReviewScreenIncrementedPhotoQuantity;
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventLevel : @2}];
-}
-
-//+ (void)trackBasketScreenHitEditItem:(id<OLPrintJob>)item inOrder:(OLPrintOrder *)printOrder applePayIsAvailable:(NSString *)applePayIsAvailable{
-//    NSString *eventName = kOLAnalyticsEventNameBasketScreenHitEditItem;
-////    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:item printOrder:printOrder extraInfo:@{kOLAnalyticsEventLevel : @2}];
-//}
-
-//+ (void)trackShippingScreenHitBackForOrder:(OLPrintOrder *)printOrder{
-////    [OLAnalytics reportAnalyticsEventToDelegate:kOLAnalyticsEventNameShippingScreenHitBack job:nil printOrder:printOrder extraInfo:@{kOLAnalyticsEventLevel : @2}];
-//}
-
-+ (void)trackReviewScreenDeletedPhotoForProductName:(NSString *)productName{
-    NSString *eventName = kOLAnalyticsEventNameReviewScreenDeletedPhoto;
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventLevel : @2}];
-}
-
-+ (void)trackProductDescriptionScreenHitBack:(NSString *)productName hidePrice:(BOOL)hidePrice{
-    NSString *eventName = kOLAnalyticsEventNameProductDescriptionScreenHitBack;
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : productName, kOLAnalyticsEventLevel : @2}];
-}
-
-+ (void)trackImagePickerScreenHitBack:(NSString *)productName{
-    NSString *eventName = kOLAnalyticsEventNameImagePickerScreenHitBack;
-//    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : nonNilStr(productName), kOLAnalyticsEventLevel : @2}];
-}
-
-+ (void)trackEditScreenDidCancel{
-//    [OLAnalytics reportAnalyticsEventToDelegate:kOLAnalyticsEventNameEditScreenFinishedEditingPhoto job:nil printOrder:nil extraInfo:@{kOLAnalyticsEventLevel : @2}];
-}
-
-+ (void)trackEditScreenButtonTapped:(NSString *)buttonName{
-    if (!buttonName){
-        return;
-    }
-//    [OLAnalytics reportAnalyticsEventToDelegate:kOLAnalyticsEventNameEditScreenButtonTapped job:nil printOrder:nil extraInfo:@{kOLAnalyticsButtonName : buttonName, kOLAnalyticsEventLevel : @2}];
-}
-
-//+ (NSMutableDictionary *)propertiesForPrintOrder:(OLPrintOrder *)printOrder {
-//    NSMutableDictionary *p = [[NSMutableDictionary alloc] init];
-//
-//    p[@"Product"] = [self listOfProductNamesForJobsInOrder:printOrder];
-//
-//    if (printOrder.proofOfPayment) {
-//        p[@"Proof of Payment"] = printOrder.proofOfPayment;
-//    }
-//
-//    if (printOrder.lastPrintSubmissionError) {
-//        p[@"Print Submission Success"] = @"False";
-//        p[@"Print Submission Error"] = nonNilStr(printOrder.lastPrintSubmissionError.localizedDescription);
-//    }
-//
-//    if (printOrder.receipt) {
-//        p[@"Print Order Id"] = printOrder.receipt;
-//        p[@"Print Submission Success"] = @"True";
-//        p[@"Print Submission Error"] = @"None";
-//    }
-//
-//    if (printOrder.promoCode) {
-//        p[@"Voucher Code"] = printOrder.promoCode;
-//    }
-//
-//    if (printOrder.email) {
-//        p[@"Shipping Email"] = printOrder.email;
-//    }
-//
-//    if (printOrder.phone) {
-//        p[@"Shipping Phone"] = printOrder.phone;
-//    }
-//
-//    if (printOrder.shippingAddress) {
-//        p[@"Shipping Recipient"] = nonNilStr(printOrder.shippingAddress.fullNameFromFirstAndLast);
-//        p[@"Shipping Line 1"] = nonNilStr(printOrder.shippingAddress.line1);
-//        p[@"Shipping Line 2"] = nonNilStr(printOrder.shippingAddress.line2);
-//        p[@"Shipping City"] = nonNilStr(printOrder.shippingAddress.city);
-//        p[@"Shipping County"] = nonNilStr(printOrder.shippingAddress.stateOrCounty);
-//        p[@"Shipping Postcode"] = nonNilStr(printOrder.shippingAddress.zipOrPostcode);
-//        p[@"Shipping Country"] = nonNilStr(printOrder.shippingAddress.country.name);
-//        p[@"Shipping Country Code2"] = nonNilStr(printOrder.shippingAddress.country.codeAlpha2);
-//        p[@"Shipping Country Code3"] = nonNilStr(printOrder.shippingAddress.country.codeAlpha3);
-//    }
-//
-//    p[@"Job Count"] = [NSString stringWithFormat:@"%lu",  (unsigned long) printOrder.jobs.count];
-//
-//    return p;
-//}
-
-//+ (NSMutableArray*) listOfProductNamesForJobsInOrder:(OLPrintOrder*) printOrder{
-//    NSMutableArray* productNames = [[NSMutableArray alloc] initWithCapacity:[printOrder.jobs count]];
-//    for (id<OLPrintJob> printJob in printOrder.jobs){
-//        id name = [OLProduct productWithTemplateId:printJob.templateId].productTemplate.name;
-//        if (name){
-//            [productNames addObject:name];
-//         }
-//    }
-//    return productNames;
-//}
 
 + (void)setExtraInfo:(NSDictionary *)info{
     extraInfo = info;
@@ -606,6 +266,13 @@ static BOOL optInToRemoteAnalytics = NO;
 
 + (NSDictionary *)extraInfo{
     return extraInfo;
+}
+
+- (void)photobookAnalyticsEventDidFireWithType:(enum AnalyticsEventType)type name:(NSString * _Nonnull)name properties:(NSDictionary<NSString *, id> * _Nonnull)properties {
+    NSMutableDictionary *dict = [properties mutableCopy];
+    dict[kOLAnalyticsEventType] = type == AnalyticsEventTypeScreenViewed ? kOLAnalyticsEventTypeScreenViewed : type == AnalyticsEventTypeAction ? kOLAnalyticsEventTypeAction : kOLAnalyticsEventTypeError;
+    dict[kOLAnalyticsEventName] = name;
+    [self.delegate logKiteAnalyticsEventWithInfo:dict];
 }
 
 @end