+++ Kite-SDK/CHANGELOG.md
@@ -8,9 +8,11 @@
 - Changed: If an order fails, the next time the SDK is launched the upload screen will show and will try to submit the order again.
 - Removed: Order history
 - Removed: Editing basket items
+- Removed: Many analytics events
 - Removed: Unused funtionality like the Promo Banner and Upsells
 - Removed: Using prerendered PDF assets and tiff images
 - Removed: Installing via Git Submodules
+- Removed: Installing via Carthage
 - Removed: Unity plugin
 - Removed: Deprecated methods.
 - Note: Theming and localization are not available in the new Photobook and checkout experiences